requirements.txt
neo4j
PyYAML
Jinja2
pydantic>=2.10.3,<3
mistralai
google-search-results
requests
beautifulsoup4
duckduckgo-search
langchain
langchain-mistralai
langgraph
wikipedia
python-dotenv
trafilatura
PyMuPDF